refactor: remove verifyInstallationStatus from CliFacade
this call had no more consumers now that CLIs handle installation status
in their own process runner error handling
