+++ src/api/aws/AwsCliFacade.ts
@@ -18,31 +18,24 @@ import { moment } from "/deps.ts";
 import { sleep } from "/promises.ts";
 import { parseJsonWithLog } from "/json.ts";
 
-import { CliFacade, CliInstallationStatus } from "../CliFacade.ts";
-
 import { ProcessResultWithOutput } from "../../process/ProcessRunnerResult.ts";
 import { ResultHandlerProcessRunnerDecorator } from "../../process/ResultHandlerProcessRunnerDecorator.ts";
 import { IProcessRunner } from "../../process/IProcessRunner.ts";
 import { AwsCliResultHandler } from "./AwsCliResultHandler.ts";
 import { CliDetector } from "../CliDetector.ts";
 
-export class AwsCliFacade implements CliFacade {
+export class AwsCliFacade {
   private readonly processRunner: IProcessRunner<ProcessResultWithOutput>;
-  private readonly detector: CliDetector;
 
   constructor(rawRunner: IProcessRunner<ProcessResultWithOutput>) {
-    this.detector = new CliDetector(rawRunner);
+    const detector = new CliDetector(rawRunner);
 
     // todo: consider wrapping the runner further, e.g. to always add --output=json so we become more independent
     // of the user's global aws cli config
     this.processRunner = new ResultHandlerProcessRunnerDecorator(
       rawRunner,
-      new AwsCliResultHandler(this.detector),
+      new AwsCliResultHandler(detector),
     );
-  }
-
-  verifyCliInstalled(): Promise<CliInstallationStatus> {
-    return this.detector.verifyCliInstalled("aws", /^aws-cli\/2\./);
   }
 
   async listProfiles(): Promise<string[]> {
