+++ src/api/gcloud/GcloudCliFacade.ts
@@ -12,7 +12,6 @@ import {
 } from "./GcpBillingExportConfig.ts";
 import { parseJsonWithLog } from "/json.ts";
 import { moment } from "/deps.ts";
-import { CliFacade, CliInstallationStatus } from "../CliFacade.ts";
 import { GcloudCliResultHandler } from "./GcloudCliResultHandler.ts";
 import { IProcessRunner } from "../../process/IProcessRunner.ts";
 import { ProcessResultWithOutput } from "../../process/ProcessRunnerResult.ts";
@@ -20,27 +19,20 @@ import { ResultHandlerProcessRunnerDecorator } from "../../process/ResultHandler
 import { CliDetector } from "../CliDetector.ts";
 import { CLI } from "../../info.ts";
 
-// todo: rename to GcloudCliFacade
-export class GcloudCliFacade implements CliFacade {
+export class GcloudCliFacade {
   private readonly processRunner: IProcessRunner<ProcessResultWithOutput>;
-  private readonly detector: CliDetector;
 
   constructor(
     rawRunner: IProcessRunner<ProcessResultWithOutput>,
     private readonly billingConfig?: GcpBillingExportConfig,
   ) {
-    this.detector = new CliDetector(rawRunner);
+    const detector = new CliDetector(rawRunner);
 
     this.processRunner = new ResultHandlerProcessRunnerDecorator(
       rawRunner,
-      new GcloudCliResultHandler(this.detector),
+      new GcloudCliResultHandler(detector),
     );
   }
-
-  verifyCliInstalled(): Promise<CliInstallationStatus> {
-    return this.detector.verifyCliInstalled("gcloud", /^Google Cloud SDK/);
-  }
-
   async configList(): Promise<Config> {
     return await this.run<Config>(["gcloud", "config", "list"]);
   }
